Moved readFile() to util/app.js
